pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "frogger"
version = "0.1.1"
description = "Repository for fast robust grasp generation"
readme = "README.md"
license = {text="Apache 2.0"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
]
dependencies = [
    "dill>=0.3.6",
    "drake>=1.21.0",
    "jax[cpu]>=0.4.1",
    "jaxlib>=0.4.1",
    "matplotlib>=3.5.2",
    "nlopt>=2.7.1",
    "numba>=0.56.4",
    "numpy>=1.23.1",
    "open3d>=0.16.0",
    "quantecon>=0.6.0",
    "scikit-image>=0.19.3",
    "scipy>=1.10.0",
    "trimesh>=4.0.0",
]


[project.optional-dependencies]
# development-only dependencies
dev = [
    "black>=23.3.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "pre-commit>=3.3.2",
]

[options]
dependency_links = [
    "https://storage.googleapis.com/jax-releases/jax_releases.html"
]

[tool.setuptools.packages.find]
include = ["frogger"]
